legacy_apps: zynqmp_r5: Add additional call restore_initial_rsc_table earlier
As the host can set the resource table before the server application takes
a snapshot, there can be error for start/stop case where there is race
condition for host-stop-remote action.

Add extra call in which server takes the resource table snapshot to
before any IPC events are in place to prevent this issue during system-init.

Signed-off-by: Ben Levinsky <ben.levinsky@amd.com>
Acked-by: Tanmay Shah <tanmay.shah@amd.com>