examples: legacy_apps: Fix Vitis build issue where libraries are queried before available
As the Vitis GUI will attempt to create application workspace without
having first built BSP, the application create step can fail.

Ensure that (a) at compile time the links are set as per convention for
Vitis GUI by PROJECT_LIB_DEPS and (b) that the values in a new domain
don't automatically break the build by removing check
for xparameters_ps.h. (b) is needed because the headers might not be at
include directory of BSP at time of application create and configure.
Acked-by: Tanmay Shah <tanmay.shah@amd.com>